backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-environ
django-cors-headers
gunicorn
dj-database-url
psycopg2-binary
Pillow
whitenoise
python-dotenv